Add handler to restart container when the configuration changed
This PR add handler + notify to restart container when you change the configuration
of the systemd service or change env vars of a container.

`register: enable_and_start` is here to avoid restart of a container during the first run.
As in case of a started state, the container is first started by the Enable and start service.
Then restarted by the handler which have been notified because the configuration "changed".

Also add `RestartSec=10s` in the template to let 10 sec before systemd try to restart a container.
This avoid to end in systemd start burst limit when a container fail.